Remove site.url
The concatenation of site.url + site.baseurl throughout the templates
seem to be causing broken links everywhere, like
“http://lanyon.getpoole.com/https://ajvondrak.github.io/soapbox/public/css/lanyon.css”

The only place I can see site.url used bare is in the 404.md, so I
removed that usage. Otherwise, I want to just leave site.url blank so
that the concatenation is a no-op that just resolves to site.baseurl.

Won't know if this works until I try it... 😓