Split URLs between site.url + site.baseurl
Almost there...now the links to the stylesheets are going to
https://ajvondrak.github.io/https://ajvondrak.github.io/soapbox/

It stands to reason that concatenating site.url + site.baseurl is
supposed to be on purpose, where the baseurl refers to the root path and
the url refers to the domain. In this light, the original 404.md makes
sense.

Guess we'll see. 🤞