Use mathjax for latex markup
Globally configures $...$ syntax for inline equations. I'll adjust this
in the future if I find myself needing to escape \$ dollar signs too
frequently.

The $$...$$ syntax wasn't being converted right from markdown locally,
and using \[...\] would eat the escaped \\ line breaks for some reason.
But wrapping it all in a <div> (or any html, really) appears to suppress
the markdown processing so that $$...$$ worked again. Bonus: it doesn't
seem to need the \\ syntax escaped.